Move commands to a separate file.
